feat: ensure unnecessary dependencies are excluded in build unless related feature is enabled
