+++ crates/rust-mcp-transport/Cargo.toml
@@ -14,17 +14,15 @@ edition = "2021"
 rust-mcp-schema = { workspace = true }
 
 async-trait = { workspace = true }
-tokio = { workspace = true }
 futures = { workspace = true }
 thiserror = { workspace = true }
 serde_json = { workspace = true }
 serde = { workspace = true }
-axum = { workspace = true }
-uuid = { workspace = true, features = ["v4"] }
 tokio-stream = { workspace = true }
-reqwest = { workspace = true, features = ["stream"] }
 bytes = { workspace = true }
 tracing = { workspace = true }
+tokio = { workspace = true }
+reqwest = { workspace = true, features = ["stream"], optional = true }
 
 [dev-dependencies]
 wiremock = "0.5"
@@ -34,10 +32,9 @@ futures = { workspace = true }
 workspace = true
 
 
-# ### FEATURES #################################################################
-# [features]
+### FEATURES #################################################################
+[features]
+default = ["stdio", "sse"] # Default features
 
-# default = ["stdio", "sse"] # Default features
-
-# stdio = []
-# sse = []
+stdio = []
+sse = ["reqwest"]
