+++ crates/rust-mcp-transport/src/utils/cancellation_token.rs
@@ -74,6 +74,7 @@ impl CancellationToken {
     ///
     /// # Returns
     /// * `bool` - True if cancellation is requested, false otherwise
+    #[allow(unused)]
     pub fn is_cancelled(&self) -> bool {
         *self.receiver.borrow()
     }
