+++ crates/rust-mcp-transport/src/utils/http_utils.rs
@@ -34,6 +34,21 @@ pub async fn http_post(
     Ok(())
 }
 
+pub fn extract_origin(url: &str) -> Option<String> {
+    // Remove the fragment first (everything after '#')
+    let url = url.split('#').next()?; // Keep only part before `#`
+
+    // Split scheme and the rest
+    let (scheme, rest) = url.split_once("://")?;
+
+    // Get host and optionally the port (before first '/')
+    let end = rest.find('/').unwrap_or(rest.len());
+    let host_port = &rest[..end];
+
+    // Reconstruct origin
+    Some(format!("{}://{}", scheme, host_port))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -146,5 +161,40 @@ mod tests {
 
         // Assert the result is an error (likely a connection error)
         assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_extract_origin_with_path() {
+        let url = "https://example.com:8080/some/path";
+        assert_eq!(
+            extract_origin(url),
+            Some("https://example.com:8080".to_string())
+        );
+    }
+
+    #[test]
+    fn test_extract_origin_without_path() {
+        let url = "https://example.com";
+        assert_eq!(extract_origin(url), Some("https://example.com".to_string()));
+    }
+
+    #[test]
+    fn test_extract_origin_with_fragment() {
+        let url = "https://example.com:8080/path#section";
+        assert_eq!(
+            extract_origin(url),
+            Some("https://example.com:8080".to_string())
+        );
+    }
+
+    #[test]
+    fn test_extract_origin_invalid_url() {
+        let url = "example.com/path";
+        assert_eq!(extract_origin(url), None);
+    }
+
+    #[test]
+    fn test_extract_origin_empty_string() {
+        assert_eq!(extract_origin(""), None);
     }
 }
